dev-ifc2tileset-agent: refactor teardown operations for unit tests
